requirements.txt
requests
psutil
python-telegram-bot>=13.0.0,<14.0.0
google-api-python-client>=2.0.0,<3.0.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
aria2p>=0.9.0,<0.15.0
python-dotenv
tenacity
python-magic
beautifulsoup4<=5.0.0
Pyrogram>=0.16.0,<0.16.10
TgCrypto
youtube_dl
megasdkrestclient>=0.1.1,<1.0.0
# python-magic-bin==0.4.14
lxml
xdcc
